Added controller for comments
Moved comments over into its own controller and updated route.
